Toggle off stopping on error while configuring harfbuzz
As it turns out, turning on BUILD_SHARED_LIBS just ends up *enforcing*
the use of the shared library in binaries that link to it, which is not
what we want.

However, turning it off causes it to emit errors about introspection,
which means GitHub Actions will stop.

To solve this for now, I'm just going to turn off stopping on errors
when harfbuzz configures, and I'll make a PR to the harfbuzz repo about
fixing this.