Split eldstar_server into "Cheats" and "TAS Safe" variants
The "Cheats" variant will in the future get dedicated features that the
"TAS Safe" variant doesn't have.
This might be stuff like patching the game camera to match Eldstar's, or
teleporting entities and Mario around.

While such functionality is obviously useful for debugging and discovery
purposes, they wouldn't be suitable during TAS recording or playback.

Some clients (such as ones dependant on actual N64 hardware) would also
be unable to observe write functionality as required by the "Cheats"
variant.

Thus, the "TAS Safe" variant is designed to be useable in all
conditions, and thus should be the default variant in most scenarios.
The "Cheats" variant is a secondary variant designed only really for
emulator-related investigation, and never for actual runs.